created MVP for current release

+++ main.py
@@ -1,9 +1,9 @@
-import shlex
-import subprocess
+import sys
 from importlib_metadata import version
 import requests
 import re
 import os
+from os.path import exists
 
 # get the most recent NodeJS version from the website
 def getNodeVersion():
@@ -28,11 +28,32 @@ def installVersion(versionNumber):
 
     if (currentVersion != versionNumber):
         # install nvm version
-        os.remove(f"node-v{versionNumber}-linux-x64.tar.gz")
+        if (exists(f"node-v{versionNumber}-linux-x64.tar.gz")):
+            os.remove(f"node-v{versionNumber}-linux-x64.tar.gz")
+
+        if (exists("/opt/nodejs")):
+            os.system(f"sudo rm -r /opt/nodejs")
+        
+        if (exists("/usr/bin/nodejs")):
+            os.system(f"sudo rm /usr/bin/nodejs")
+
         os.system(f"wget https://nodejs.org/dist/v{versionNumber}/node-v{versionNumber}-linux-x64.tar.gz")
         os.system(f"tar -xvzf node-v{versionNumber}-linux-x64.tar.gz")
+        os.system(f"sudo mv node-v{versionNumber}-linux-x64 /opt/nodejs")
+        os.system("sudo ln -s /opt/nodejs/bin/node /usr/bin/nodejs")
+        # os.system("sudo cp /opt/nodejs/bin/node /usr/bin/nodejs")
     else:
         print("NodeJS is currently up to date")
+
+def printHelp():
+    helpPage = """
+    Usage: python3 main.py [version]
+    Options for version: LTS
+                         latest
+                         [version number]
+    """
+
+    print(helpPage)
 
 if __name__ == "__main__":
     nodeVersion = getNodeVersion()
@@ -44,3 +65,8 @@ if __name__ == "__main__":
     
     print(f"Installing NodeJS Version {nodeVersion}")
     installVersion(nodeVersion)
+
+    print("")
+    print("NodeJS installation complete")
+    print("Installed as nodejs, check by running")
+    print("$ nodejs --version")
